+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New Features
 
 - Expanded RTCConfiguration support. We now support `iceTransportPolicy`,
   `bundlePolicy`, `rtcpMuxPolicy`, and `iceCandidatePoolSize`.
+- Added support for `getConfiguration` and `setConfiguration`.
 - Added support for a non-standard RTCConfiguration attribute, `portRange`. You
   can use this to constrain the `min` and `max` ports to be used when creating
   ICE transports. For example,
@@ -14,6 +15,8 @@ New Features
   const { RTCPeerConnection } = require('wrtc');
   const pc = new RTCPeerConnection({ portRange: { min: 0, max: 65535 } });
   ```
+
+  `portRange` cannot be updated by `setConfiguration`.
 
 0.0.67
 ======